docs: added some docs, little code refactoring

+++ core/src/com/knightwars/game/GameManager.java
@@ -6,9 +6,10 @@ public class GameManager {
 
     public GameManager() { this.game = new KnightWarsGame(); }
 
-    /* Update the game by a certain value */
-    public void update(int dt) {}
+    /** Update the game
+     * @param dt time parameter
+     */
+    public void update(float dt) {}
 
-    /* KnightWarsGame getter */
     public KnightWarsGame getKnightWarsGame() { return this.game; }
 }
